integration: Refactors container_update_resources_tests.go
Moves TestUpdateContainerResources_MemorySwap and associated functions
to container_update_resources_linux_test.go. They are Linux-specific,
requiring cgroups. We're doing this in order to allow the other tests
to have Windows support afterwards.
